Minor changes, fix typos
